// interfaces/IMessage.ts
export interface IMessage<T> {
  event: SocketEvents;
  data: T;
}

export enum SocketEvents {
  GET_ROOMS = "GET_ROOMS",
  ADD_ROOM = "ADD_ROOM",
  REMOVE_ROOM = "REMOVE_ROOM",
  CREATE_ROOM = "CREATE_ROOM",
  CREATE_ROOM_ERROR = "CREATE_ROOM_ERROR",
  CREATE_ROOM_SUCCESS = "CREATE_ROOM_SUCCESS",
  JOIN_ROOM = "JOIN_ROOM",
  JOIN_ROOM_ERROR = "JOIN_ROOM_ERROR",
  JOIN_ROOM_SUCCESS = "JOIN_ROOM_SUCCESS",
  CAN_PLAY = "CAN_PLAY",
  CAN_PLAY_ERROR = "CAN_PLAY_ERROR",
  CAN_PLAY_SUCCESS = "CAN_PLAY_SUCCESS",
  PICK = "PICK",
  SYMBOL = "SYMBOL",
  TURN_ERROR = "TURN_ERROR",
  TURN = "TURN",
  UNKNOWN_EVENT = "UNKNOWN_EVENT",
  PONG = "PONG",
  PING = "PING",
  GAME_RESULT = "GAME_RESULT",
}